\name{liner}
\alias{liner}
\docType{data}
\title{
The marine engine cylinder liner data from Giorgio et al. (2012)
}
\description{
The degradation data measured on 32 marine engine cylinder liners from January 1999 to August 2006. See Giorgio et al. (2012) for details.
}
\usage{data("liner")}
\format{
  A data frame with 64 observations on the following 3 variables.
  \describe{
    \item{\code{ID}}{liner ID.}
    \item{\code{Timepoints}}{the time of measurement (in years).}
    \item{\code{Measurements}}{the measurement (in mm).}
  }
}
\source{
Giorgio M., Guida M., and Pulcini G.(2012). A State-Dependent Wear Model with an Application to Marine Engine Cylinder Liners. Technometrics, 52(2), 172-187.
}
\references{
Giorgio M., Guida M., and Pulcini G.(2012). A State-Dependent Wear Model with an Application to Marine Engine Cylinder Liners. Technometrics, 52(2), 172-187.
}
\keyword{datasets}
